src/main/java/jpabook/jpashop/MemberRepository.java
package jpabook.jpashop;

import org.springframework.stereotype.Repository;

import javax.persistence.EntityManager;
import javax.persistence.PersistenceContext;

@Repository
public class MemberRepository {

    // Shift + Ctrl + T 테스트 코드 생성

    // 스프링 부트가 EntityManager (Raw JPA)를 주입
    @PersistenceContext
    private EntityManager em;

    public Long save(Member member){
        em.persist(member);
        return member.getId();
    }

    public Member find(Long id){
        return em.find(Member.class , id);
    }

}
